new package at.ac.tuwien.kr.alpha.api for classes making up the Alpha's public API, moved Alpha there
